+++ modules/apps/layout/layout-reports-web/src/main/resources/META-INF/resources/js/components/render_times/Filter.tsx
@@ -15,7 +15,7 @@ import {
 	FILTER_TYPE_NAMES,
 	FRAGMENT_FILTERS,
 	FragmentFilter,
-} from '../../constants/fragments';
+} from '../../constants/Fragment';
 
 interface Props {
 	filters: FragmentFilter;
