+++ modules/apps/layout/layout-reports-web/src/main/resources/META-INF/resources/js/components/render_times/ResultsBar.tsx
@@ -20,7 +20,7 @@ import {
 	FILTER_NAMES,
 	Fragment,
 	FragmentFilter,
-} from '../../constants/fragments';
+} from '../../constants/Fragment';
 
 interface PropsResultsBar {
 	className: string;
