+++ modules/apps/layout/layout-reports-web/src/main/resources/META-INF/resources/js/components/render_times/getFragmentsByFilterValue.ts
@@ -3,7 +3,7 @@
  * SPDX-License-Identifier: LGPL-2.1-or-later OR LicenseRef-Liferay-DXP-EULA-2.0.0-2023-06
  */
 
-import {Fragment, FragmentFilter} from '../../constants/fragments';
+import {Fragment, FragmentFilter} from '../../constants/Fragment';
 
 export default function getFragmentsByFilterValue(
 	filters: FragmentFilter,
