+++ modules/apps/layout/layout-reports-web/src/main/resources/META-INF/resources/js/constants/Fragment.ts
